+++ .vscode/settings.json
@@ -1,1 +1,4 @@
-{"editor.formatOnSave": true}
+{
+  "editor.formatOnSave": true,
+  "biome.enabled": false
+}
