+++ new/src/lib/assertBlockKey.ts
