+++ new/src/lib/defaultComponents/DefaultListItem.svelte
