+++ new/src/lib/renderers/RenderListItem.svelte
