+++ new/src/lib/renderers/RenderText.svelte
