--- vitest.config.ts
@@ -1,10 +1,0 @@
-import {defineConfig} from 'vite'
-import {svelte} from '@sveltejs/vite-plugin-svelte'
-
-export default defineConfig({
-  plugins: [svelte({hot: !process.env.VITEST})],
-  test: {
-    globals: true,
-    environment: 'jsdom'
-  }
-} as any)
